+++ changelog.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Breaking
+- build Metadata class from protobuf definition
+
 ## [0.1.0] - 2024-12-21
 
 ### Added